subaru.project/requirements.txt
Django>=5.2,<6.1
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
drf-yasg>=1.21,<2.0
Pillow>=11.0,<12.0
